% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{AAcenter}
\alias{AAcenter}
\title{Center AA data}
\usage{
AAcenter(AAdat, SampleIndices, AAIndices = NA)
}
\arguments{
\item{AAdat}{An Autoantigen Array object}

\item{SampleIndices}{The indices of the samples to be centered.}

\item{AAIndices}{The indices of the autoantigens used to center the data. If NA, then it is al features}
}
\description{
Center AA data
}

