% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{AAplotByMean}
\alias{AAplotByMean}
\title{Plot AA data by mean}
\usage{
AAplotByMean(AAdat, SampleIndices, clrs = 1, pchs = 3, MeanSubset = NA,
  useMat = "log", cex = 0.5, ...)
}
\arguments{
\item{AAdat}{An Autoantigen Array object}

\item{SampleIndices}{The indices of the samples to be selected.}

\item{clrs}{colors for the samples. If length 1 then they are all that color}

\item{pchs}{point type for the samples. If length 1 then they are all that point type}

\item{MeanSubset}{subset index of samples by which to order the means of the features}

\item{useMat}{which data matrix to use for the data points. Currently, "log", "raw", and "resid" are available}

\item{cex}{the size of the plotted points}
}
\description{
Plot AA data by mean
}

