% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{AAplotDens}
\alias{AAplotDens}
\title{Plot Densities of AA data}
\usage{
AAplotDens(AAdat, SampleIndices, clrs = 1, useMat = "log", xlims = NA,
  returnModes = T, ...)
}
\arguments{
\item{AAdat}{An Autoantigen Array object}

\item{SampleIndices}{The indices of the samples to be selected.}

\item{clrs}{colors for the samples. If length 1 then they are all that color}

\item{useMat}{which data matrix to use for the data points. Currently, "log", "raw", and "resid" are available}

\item{xlims}{the plotting range for the figure. A value of NA will ustilize the range of the data}

\item{returnModes}{flag to return estimates of the density modes}
}
\description{
Plot Densities of AA data
}

