% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{ExtractFromGenePix}
\alias{ExtractFromGenePix}
\title{Extract Data From GenePix List Object}
\usage{
ExtractFromGenePix(GPlist, SignalName, SampleIndices)
}
\arguments{
\item{GPlist}{GenePix data list.}

\item{SignalName}{Name of the GenePix column that will be used to quantify the signal.}

\item{SampleIndices}{The indices of the samples to be selected.}
}
\description{
Extract Data From GenePix List Object
}

