% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{LogTransform}
\alias{LogTransform}
\title{Log Transform the Data}
\usage{
LogTransform(AAdat, const = 1, delta = NA, offsets = NA)
}
\arguments{
\item{AAdat}{An Autoantigen Array object}

\item{const}{The consant to include in the log transfom. cnst=NA means no constant}

\item{delta}{option to specify delta exactly}

\item{offsets}{option to supply sample specific offsets}
}
\description{
Log Transform the Data
}

