% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{RowTriMeans}
\alias{RowTriMeans}
\title{Calculate Tukey Tri-means for the Rows of a matrix}
\usage{
RowTriMeans(X, type = 8)
}
\arguments{
\item{X}{a matrix of values}

\item{type}{the method of quantile estimation that will be employed}
}
\description{
Calculate Tukey Tri-means for the Rows of a matrix
}

