% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromGenePix.R
\name{TriMean}
\alias{TriMean}
\title{Calculate Tukey Tri-mean}
\usage{
TriMean(x, type = 8)
}
\arguments{
\item{x}{a vector of values}

\item{type}{the method of quantile estimation that will be employed}
}
\description{
Calculate Tukey Tri-mean
}

